Added network class for gripper.

+++ src/robot_network.cpp
@@ -15,7 +15,8 @@ RobotNetwork::RobotNetwork(const std::string& franka_address,
 
 research_interface::robot::RobotState RobotNetwork::udpReadRobotState() try {
   std::array<uint8_t, sizeof(research_interface::robot::RobotState)> buffer;
-  int bytes_received = udp_socket_.receiveFrom(buffer.data(), buffer.size(), udp_server_address_);
+  int bytes_received =
+      udp_socket_.receiveFrom(buffer.data(), static_cast<int>(buffer.size()), udp_server_address_);
   if (bytes_received != buffer.size()) {
     throw ProtocolException("libfranka robot: incorrect object size");
   }
